+++ src/Phinx/Db/Adapter/AdapterInterface.php
@@ -81,8 +81,9 @@ interface AdapterInterface
     public function getVersions();
 
     /**
-     * Get all migration log entries, indexed by version number.
-     *
+     * Get all migration log entries, indexed by version creation time and sorted ascendingly by the configuration's 
+     * version order option
+     * 
      * @return array
      */
     public function getVersionLog();
@@ -330,6 +331,15 @@ interface AdapterInterface
      */
     public function dropTable($tableName);
 
+
+    /**
+     * Truncates the specified table
+     *
+     * @param string $tableName
+     * @return void
+     */
+    public function truncateTable($tableName);
+
     /**
      * Returns table columns
      *
@@ -479,7 +489,7 @@ interface AdapterInterface
      *
      * @param string $type
      * @param integer $limit
-     * @return string
+     * @return string[]
      */
     public function getSqlType($type, $limit = null);
 
@@ -507,4 +517,13 @@ interface AdapterInterface
      * @return void
      */
     public function dropDatabase($name);
+
+    /**
+     * Cast a value to a boolean appropriate for the adapter.
+     *
+     * @param mixed $value The value to be cast
+     *
+     * @return mixed
+     */
+    public function castToBool($value);
 }