+++ src/Phinx/Db/Adapter/AdapterWrapper.php
@@ -148,7 +148,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function connect()
     {
-        return $this->getAdapter()->connect();
+        $this->getAdapter()->connect();
     }
 
     /**
@@ -156,7 +156,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function disconnect()
     {
-        return $this->getAdapter()->disconnect();
+        $this->getAdapter()->disconnect();
     }
 
     /**
@@ -180,7 +180,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function insert(Table $table, $row)
     {
-        return $this->getAdapter()->insert($table, $row);
+        $this->getAdapter()->insert($table, $row);
     }
 
     /**
@@ -254,7 +254,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function createSchemaTable()
     {
-        return $this->getAdapter()->createSchemaTable();
+        $this->getAdapter()->createSchemaTable();
     }
 
     /**
@@ -286,7 +286,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function beginTransaction()
     {
-        return $this->getAdapter()->beginTransaction();
+        $this->getAdapter()->beginTransaction();
     }
 
     /**
@@ -294,7 +294,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function commitTransaction()
     {
-        return $this->getAdapter()->commitTransaction();
+        $this->getAdapter()->commitTransaction();
     }
 
     /**
@@ -302,7 +302,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function rollbackTransaction()
     {
-        return $this->getAdapter()->rollbackTransaction();
+        $this->getAdapter()->rollbackTransaction();
     }
 
     /**
@@ -334,7 +334,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function createTable(Table $table)
     {
-        return $this->getAdapter()->createTable($table);
+        $this->getAdapter()->createTable($table);
     }
 
     /**
@@ -342,7 +342,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function renameTable($tableName, $newTableName)
     {
-        return $this->getAdapter()->renameTable($tableName, $newTableName);
+        $this->getAdapter()->renameTable($tableName, $newTableName);
     }
 
     /**
@@ -350,7 +350,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function dropTable($tableName)
     {
-        return $this->getAdapter()->dropTable($tableName);
+        $this->getAdapter()->dropTable($tableName);
     }
 
     /**
@@ -374,7 +374,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function addColumn(Table $table, Column $column)
     {
-        return $this->getAdapter()->addColumn($table, $column);
+        $this->getAdapter()->addColumn($table, $column);
     }
 
     /**
@@ -382,7 +382,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function renameColumn($tableName, $columnName, $newColumnName)
     {
-        return $this->getAdapter()->renameColumn($tableName, $columnName, $newColumnName);
+        $this->getAdapter()->renameColumn($tableName, $columnName, $newColumnName);
     }
 
     /**
@@ -398,7 +398,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function dropColumn($tableName, $columnName)
     {
-        return $this->getAdapter()->dropColumn($tableName, $columnName);
+        $this->getAdapter()->dropColumn($tableName, $columnName);
     }
 
     /**
@@ -422,15 +422,15 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function addIndex(Table $table, Index $index)
     {
-        return $this->getAdapter()->addIndex($table, $index);
+        $this->getAdapter()->addIndex($table, $index);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function dropIndex($tableName, $columns, $options = array())
+    public function dropIndex($tableName, $columns)
     {
-        return $this->getAdapter()->dropIndex($tableName, $columns, $options);
+        $this->getAdapter()->dropIndex($tableName, $columns);
     }
 
     /**
@@ -438,7 +438,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function dropIndexByName($tableName, $indexName)
     {
-        return $this->getAdapter()->dropIndexByName($tableName, $indexName);
+        $this->getAdapter()->dropIndexByName($tableName, $indexName);
     }
 
     /**
@@ -454,7 +454,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function addForeignKey(Table $table, ForeignKey $foreignKey)
     {
-        return $this->getAdapter()->addForeignKey($table, $foreignKey);
+        $this->getAdapter()->addForeignKey($table, $foreignKey);
     }
 
     /**
@@ -462,7 +462,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function dropForeignKey($tableName, $columns, $constraint = null)
     {
-        return $this->getAdapter()->dropForeignKey($tableName, $columns, $constraint);
+        $this->getAdapter()->dropForeignKey($tableName, $columns, $constraint);
     }
 
     /**
@@ -478,7 +478,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function createDatabase($name, $options = array())
     {
-        return $this->getAdapter()->createDatabase($name, $options);
+        $this->getAdapter()->createDatabase($name, $options);
     }
 
     /**
@@ -486,7 +486,7 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function hasDatabase($name)
     {
-        return $this->getAdapter()->hasDatabase($name);
+        $this->getAdapter()->hasDatabase($name);
     }
 
     /**
@@ -494,11 +494,11 @@ abstract class AdapterWrapper implements AdapterInterface, WrapperInterface
      */
     public function dropDatabase($name)
     {
-        return $this->getAdapter()->dropDatabase($name);
+        $this->getAdapter()->dropDatabase($name);
     }
 
     /**
-     * @inheritDoc
+     * {@inheritdoc}
      */
     public function castToBool($value)
     {