+++ src/Phinx/Db/Adapter/TablePrefixAdapter.php
@@ -74,7 +74,7 @@ class TablePrefixAdapter extends AdapterWrapper
             $adapterReferenceTable->setName($adapterReferenceTableName);
         }
 
-        return parent::createTable($adapterTable);
+        parent::createTable($adapterTable);
     }
 
     /**
@@ -84,7 +84,7 @@ class TablePrefixAdapter extends AdapterWrapper
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
         $adapterNewTableName = $this->getAdapterTableName($newTableName);
-        return parent::renameTable($adapterTableName, $adapterNewTableName);
+        parent::renameTable($adapterTableName, $adapterNewTableName);
     }
 
     /**
@@ -93,7 +93,16 @@ class TablePrefixAdapter extends AdapterWrapper
     public function dropTable($tableName)
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
-        return parent::dropTable($adapterTableName);
+        parent::dropTable($adapterTableName);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function truncateTable($tableName)
+    {
+        $adapterTableName = $this->getAdapterTableName($tableName);
+        return parent::truncateTable($adapterTableName);
     }
 
     /**
@@ -122,7 +131,7 @@ class TablePrefixAdapter extends AdapterWrapper
         $adapterTable = clone $table;
         $adapterTableName = $this->getAdapterTableName($table->getName());
         $adapterTable->setName($adapterTableName);
-        return parent::addColumn($adapterTable, $column);
+        parent::addColumn($adapterTable, $column);
     }
 
     /**
@@ -131,7 +140,7 @@ class TablePrefixAdapter extends AdapterWrapper
     public function renameColumn($tableName, $columnName, $newColumnName)
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
-        return parent::renameColumn($adapterTableName, $columnName, $newColumnName);
+        parent::renameColumn($adapterTableName, $columnName, $newColumnName);
     }
 
     /**
@@ -149,7 +158,7 @@ class TablePrefixAdapter extends AdapterWrapper
     public function dropColumn($tableName, $columnName)
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
-        return parent::dropColumn($adapterTableName, $columnName);
+        parent::dropColumn($adapterTableName, $columnName);
     }
 
     /**
@@ -178,16 +187,16 @@ class TablePrefixAdapter extends AdapterWrapper
         $adapterTable = clone $table;
         $adapterTableName = $this->getAdapterTableName($table->getName());
         $adapterTable->setName($adapterTableName);
-        return parent::addIndex($adapterTable, $index);
+        parent::addIndex($adapterTable, $index);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function dropIndex($tableName, $columns, $options = array())
+    public function dropIndex($tableName, $columns)
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
-        return parent::dropIndex($adapterTableName, $columns, $options);
+        parent::dropIndex($adapterTableName, $columns);
     }
 
     /**
@@ -196,7 +205,7 @@ class TablePrefixAdapter extends AdapterWrapper
     public function dropIndexByName($tableName, $indexName)
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
-        return parent::dropIndexByName($adapterTableName, $indexName);
+        parent::dropIndexByName($adapterTableName, $indexName);
     }
 
     /**
@@ -216,7 +225,7 @@ class TablePrefixAdapter extends AdapterWrapper
         $adapterTable = clone $table;
         $adapterTableName = $this->getAdapterTableName($table->getName());
         $adapterTable->setName($adapterTableName);
-        return parent::addForeignKey($adapterTable, $foreignKey);
+        parent::addForeignKey($adapterTable, $foreignKey);
     }
 
     /**
@@ -225,7 +234,7 @@ class TablePrefixAdapter extends AdapterWrapper
     public function dropForeignKey($tableName, $columns, $constraint = null)
     {
         $adapterTableName = $this->getAdapterTableName($tableName);
-        return parent::dropForeignKey($adapterTableName, $columns, $constraint);
+        parent::dropForeignKey($adapterTableName, $columns, $constraint);
     }
 
     /**
@@ -236,9 +245,9 @@ class TablePrefixAdapter extends AdapterWrapper
         $adapterTable = clone $table;
         $adapterTableName = $this->getAdapterTableName($table->getName());
         $adapterTable->setName($adapterTableName);
-        return parent::insert($adapterTable, $row);
+        parent::insert($adapterTable, $row);
     }
-    
+
     /**
      * Gets the table prefix.
      *