+++ src/Phinx/Seed/SeedInterface.php
@@ -30,6 +30,7 @@ namespace Phinx\Seed;
 
 use Phinx\Db\Adapter\AdapterInterface;
 use Phinx\Db\Table;
+use Phinx\Migration\MigrationInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -56,7 +57,7 @@ interface SeedInterface
      * Sets the database adapter.
      *
      * @param AdapterInterface $adapter Database Adapter
-     * @return MigrationInterface
+     * @return SeedInterface
      */
     public function setAdapter(AdapterInterface $adapter);
 
@@ -71,7 +72,7 @@ interface SeedInterface
      * Sets the input object to be used in migration object
      *
      * @param InputInterface $input
-     * @return MigrationInterface
+     * @return SeedInterface
      */
     public function setInput(InputInterface $input);
 
@@ -86,7 +87,7 @@ interface SeedInterface
      * Sets the output object to be used in migration object
      *
      * @param OutputInterface $output
-     * @return MigrationInterface
+     * @return SeedInterface
      */
     public function setOutput(OutputInterface $output);